TST move test_model_pipeline_same_dense_and_sparse to test_common.py
