+++ sklearn/linear_model/tests/test_coordinate_descent.py
@@ -9,7 +9,7 @@ import numpy as np
 import pytest
 from scipy import interpolate, sparse
 
-from sklearn.base import clone, config_context, is_classifier
+from sklearn.base import clone, config_context
 from sklearn.datasets import load_diabetes, make_regression
 from sklearn.exceptions import ConvergenceWarning
 from sklearn.linear_model import (
@@ -19,7 +19,6 @@ from sklearn.linear_model import (
     LassoCV,
     LassoLars,
     LassoLarsCV,
-    LinearRegression,
     MultiTaskElasticNet,
     MultiTaskElasticNetCV,
     MultiTaskLasso,
@@ -356,56 +355,6 @@ def _scale_alpha_inplace(estimator, n_samples):
             raise NotImplementedError
 
     estimator.set_params(alpha=alpha)
-
-
-@pytest.mark.filterwarnings("ignore::sklearn.exceptions.ConvergenceWarning")
-@pytest.mark.parametrize(
-    "LinearModel, params",
-    [
-        (Lasso, {"tol": 1e-16, "alpha": 0.1}),
-        (LassoCV, {"tol": 1e-16}),
-        (ElasticNetCV, {}),
-        (RidgeClassifier, {"solver": "sparse_cg", "alpha": 0.1}),
-        (ElasticNet, {"tol": 1e-16, "l1_ratio": 1, "alpha": 0.01}),
-        (ElasticNet, {"tol": 1e-16, "l1_ratio": 0, "alpha": 0.01}),
-        (Ridge, {"solver": "sparse_cg", "tol": 1e-12, "alpha": 0.1}),
-        (LinearRegression, {}),
-        (RidgeCV, {}),
-        (RidgeClassifierCV, {}),
-    ],
-)
-@pytest.mark.parametrize("csr_container", CSR_CONTAINERS)
-def test_model_pipeline_same_dense_and_sparse(LinearModel, params, csr_container):
-    # Test that linear model preceded by StandardScaler in the pipeline and
-    # with normalize set to False gives the same y_pred and the same .coef_
-    # given X sparse or dense
-
-    model_dense = make_pipeline(StandardScaler(with_mean=False), LinearModel(**params))
-
-    model_sparse = make_pipeline(StandardScaler(with_mean=False), LinearModel(**params))
-
-    # prepare the data
-    rng = np.random.RandomState(0)
-    n_samples = 200
-    n_features = 2
-    X = rng.randn(n_samples, n_features)
-    X[X < 0.1] = 0.0
-
-    X_sparse = csr_container(X)
-    y = rng.rand(n_samples)
-
-    if is_classifier(model_dense):
-        y = np.sign(y)
-
-    model_dense.fit(X, y)
-    model_sparse.fit(X_sparse, y)
-
-    assert_allclose(model_sparse[1].coef_, model_dense[1].coef_)
-    y_pred_dense = model_dense.predict(X)
-    y_pred_sparse = model_sparse.predict(X_sparse)
-    assert_allclose(y_pred_dense, y_pred_sparse)
-
-    assert_allclose(model_dense[1].intercept_, model_sparse[1].intercept_)
 
 
 def test_lasso_path_return_models_vs_new_return_gives_same_coefficients():
