+++ sklearn/utils/tests/test_pprint.py
@@ -406,7 +406,7 @@ def test_gridsearch_pipeline(print_changed_only_false):
             "classify__C": C_OPTIONS,
         },
     ]
-    gspipline = GridSearchCV(pipeline, cv=3, n_jobs=1, param_grid=param_grid)
+    gspipeline = GridSearchCV(pipeline, cv=3, n_jobs=1, param_grid=param_grid)
     expected = """
 GridSearchCV(cv=3, error_score='raise-deprecating',
              estimator=Pipeline(memory=None,
@@ -447,7 +447,7 @@ GridSearchCV(cv=3, error_score='raise-deprecating',
              scoring=None, verbose=0)"""  # noqa: E501
 
     expected = expected[1:]  # remove first \n
-    repr_ = pp.pformat(gspipline)
+    repr_ = pp.pformat(gspipeline)
     # Remove address of '<function chi2 at 0x.....>' for reproducibility
     repr_ = re.sub("function chi2 at 0x.*>", "function chi2 at some_address>", repr_)
     assert repr_ == expected
